+++ databend-jdbc/src/main/java/com/databend/jdbc/internal/http/DatabendPresignClient.java
@@ -1,4 +1,4 @@
-package com.databend.jdbc.cloud;
+package com.databend.jdbc.internal.http;
 
 import okhttp3.Headers;
